Move SSO handlers to ssohandler package
