+++ pkg/auth/handler/ssohandler/login_auth_url.go
@@ -1,4 +1,4 @@
-package handler
+package ssohandler
 
 import (
 	"encoding/json"
